requirements.txt
docutils~=0.18.1
Sphinx~=5.0.2
pandas>=1.3.0
numpy>=1.19.0
h5py>=3.1.0
torch>=1.10.0
six~=1.16.0
pyfaidx>=0.7.0
pytabix
PyYAML>=5.1
sklearn~=0.0
scikit-learn>=1.0.0
click>=8.0.0
scipy>=1.7.0
tensorflow>=2.9.0
matplotlib>=2.2.3
plotly>=5.0.0
seaborn>=0.11.0
Cython>=0.27.3
setuptools>=42.0.0
Jinja2>=3.0.0
statsmodels>=0.13.0
bioPython>=1.73
pydantic>=1.9.0
torchinfo